server/requirements.txt
fastapi
uvicorn
requests
pandas
langchain
openai
pinecone-client
python-dotenv
statsmodels
